- from corresponding shift find events in interviews calendar - parse through events to find tech/BPM interviews and fit interviews - extract name, email, start time for future use with hackerRank invites, email invites
